refactor: 페이지 디렉토리 구조 변경
