+++ src/index.tsx
@@ -47,8 +47,8 @@ import AnnouncementDetail from 'pages/Setting/Announcement/AnnouncementDetail';
 import FaqList from 'pages/Setting/Faq/FaqList';
 import FaqDetail from 'pages/Setting/Faq/FaqDetail';
 import EditUserProfile from 'pages/Setting/EditUserProfile';
-import Community from 'pages/Community';
-import CommunityWrite from 'pages/CommunityWrite';
+import Community from 'pages/Community/CommunityMain';
+import CommunityWrite from 'pages/Community/CommunityWrite';
 // 모바일 100vh 세팅
 
 let vh = window.innerHeight * 0.01;
