+++ src/pages/Community/CommunityDetail/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const CommunityDetail = () => {
+  return <></>;
+};
+
+export default CommunityDetail;
