+++ src/pages/Community/CommunityMain/CommunityHeader/PostTypeSelector.tsx
