+++ src/pages/Community/CommunityMain/CommunityHeader/SearchBar.tsx
