+++ src/pages/Community/CommunityMain/types.ts
@@ -1,0 +1,7 @@
+export interface PostList {
+  searchContent: string;
+  offset: number;
+  limit: number;
+  postType: 'INFORMATION_SHARE' | 'GARDEN_SHOWCASE' | 'QUESTION' | 'ETC';
+  orderBy: 'COMMENT_COUNT' | 'RECENT_DATE' | 'LIKE_COUNT' | 'OLDER_DATE';
+}
